SS-3902,Refactor PEER environmnet PEK creation

+++ management/server/core/security-manager/security-manager-api/src/main/java/io/subutai/core/security/api/model/SecurityKeyIdentity.java
@@ -34,13 +34,13 @@ public interface SecurityKeyIdentity
     /********************************
      *
      */
-    public short getType();
+    public int getType();
 
 
     /********************************
      *
      */
-    public void setType( final short type );
+    public void setType( final int type );
 
 
     /********************************
